Fix soundness issues with clipboard implementation
Note that the implementation of copying to the clipboard relies on
forking the process. The child process is then responsible for managing
the clipboard, and the parent process is responsible for managing the
pass file.

Previously, we would stop tracking if the process was forked (and if it
is therefore the parent or the child) once an error had occurred. This
led to issues like both processes attempting to recover the backup
simultaneously on error. This has now been fixed - if the process is
forked, only the parent process will manage the backup.

Additionally, if the current process is the child process, we would
mistakenly disable memory protections early. This has now been fixed
such that memory protections are reapplied to the child process and last
until all sensitive data is dropped.